# Improve accuracy of pow function with new implementation