--- a/tt_metal/hw/ckernels/wormhole_b0/metal/llk_api/llk_sfpu/ckernel_sfpu_binary.h
+++ b/tt_metal/hw/ckernels/wormhole_b0/metal/llk_api/llk_sfpu/ckernel_sfpu_binary.h
@@ -72,8 +72,8 @@ sfpi_inline sfpi::vFloat _sfpu_binary_power_(sfpi::vFloat base, sfpi::vFloat pow
 
     // Step 1: Compute log2(base)
     // Normalize base to calculation range
-    sfpi::vFloat x = setsgn(base, 0);  // set base as positive
-    x = sfpi::setexp(x, 127);          // set exp to exp bias (put base in range of 1-2)
+    sfpi::vFloat absbase = setsgn(base, 0);       // set base as positive
+    sfpi::vFloat x = sfpi::setexp(absbase, 127);  // set exp to exp bias (put base in range of 1-2)
 
     // 3rd order polynomial approx - determined using rminimax over [1,2]
     sfpi::vFloat series_result = x * (x * (x * 0x2.44734p-4f - 0xd.e712ap-4f) + 0x2.4f5388p+0f) - 0x1.952992p+0f;
@@ -122,7 +122,7 @@ sfpi_inline sfpi::vFloat _sfpu_binary_power_(sfpi::vFloat base, sfpi::vFloat pow
     sfpi::vFloat pow_rounded = sfpi::int32_to_float(pow_int, 0);
 
     // Division by 0 when base is 0 and pow is negative => set to NaN
-    v_if((base == 0.f) && pow < 0.f) {
+    v_if((absbase == 0.f) && pow < 0.f) {
         y = sfpi::vConstFloatPrgm2;  // negative powers of 0 are NaN, e.g. pow(0, -1.5)
     }
     v_endif;