--- a/tt_metal/hw/ckernels/wormhole_b0/metal/llk_api/llk_sfpu/ckernel_sfpu_binary.h
+++ b/tt_metal/hw/ckernels/wormhole_b0/metal/llk_api/llk_sfpu/ckernel_sfpu_binary.h
@@ -38,7 +38,39 @@ sfpi_inline sfpi::vInt _float_to_int32_positive_(sfpi::vFloat in) {
     return result;
 }
 
+/**
+ * @brief Computes base raised to the power of pow (base**pow)
+ *
+ * This function implements binary exponentiation using a polynomial approximation algorithm
+ * based on "Simple Multiple Precision Algorithms for Exponential Functions [Tips & Tricks]"
+ * by Moroz et al. 2022 (https://doi.org/10.1109/MSP.2022.3157460).
+ * More specifically, it is the implementation of the `exp_21f` algorithm described in Section 5
+ *
+ * @param base The base value (sfpi::vFloat vector), can be any floating point number
+ * @param pow The exponent/power value (sfpi::vFloat vector), can be any floating point number
+ *
+ * @return sfpi::vFloat Result of base**pow
+ *
+ * Special Cases:
+ * - base = 0, pow < 0: Returns NaN (undefined)
+ * - base < 0, pow = integer: Returns proper signed result (negative if odd power)
+ * - base < 0, pow = non-integer: Returns NaN (complex result)
+ * - Overflow/underflow: Clamped to appropriate limits
+ *
+ * @note This functions assumes that the programmable constants are set to the following values:
+ * - vConstFloatPrgm0 = 1.4426950408889634f;
+ * - vConstFloatPrgm1 = -127.0f;
+ * - vConstFloatPrgm2 = std::numeric_limits<float>::quiet_NaN();
+ *
+ * @see Moroz et al. 2022 - "Simple Multiple Precision Algorithms for Exponential Functions"
+ *      ( https://doi.org/10.1109/MSP.2022.3157460 )
+ */
 sfpi_inline sfpi::vFloat _sfpu_binary_power_(sfpi::vFloat base, sfpi::vFloat pow) {
+    // THe algorithm works in two steps:
+    // 1) Compute log2(base)
+    // 2) Compute base**pow = 2**(pow * log2(base))
+
+    // Step 1: Compute log2(base)
     // Normalize base to calculation range
     sfpi::vFloat x = setsgn(base, 0);  // set base as positive
     x = sfpi::setexp(x, 127);          // set exp to exp bias (put base in range of 1-2)
@@ -56,27 +88,30 @@ sfpi_inline sfpi::vFloat _sfpu_binary_power_(sfpi::vFloat base, sfpi::vFloat pow
     const sfpi::vFloat vConst1Ln2 = sfpi::vConstFloatPrgm0;        // 1.4426950408889634f;
     sfpi::vFloat log2_result = expf + series_result * vConst1Ln2;  // exp correction: ln(1+x) + exp*ln(2)
 
+    // Step 2: Compute base**pow = 2**(pow * log2(base))
+    // If (base, exponent) => (0, +inf) or (base, exponent) => (N, -inf) then ooutput should be 0
+    // However, intermediary values can overflow, which leads to output increasing again instead of
+    // staying at 0.
+    // This overflows happens when zff < -127. Therefore, we clamp zff to -127.
     sfpi::vFloat zff = pow * log2_result;
     const sfpi::vFloat low_threshold = sfpi::vConstFloatPrgm1;
-    v_if(zff < low_threshold)  // -126.99999237060546875
-    {
-        zff = low_threshold;
-    }
+    v_if(zff < low_threshold) { zff = low_threshold; }
     v_endif;
 
-    zff = addexp(zff, 23);                                                 // * 2**23 (Mn)
+    zff = addexp(zff, 23);                                                     // * 2**23 (Mn)
     sfpi::vInt z = _float_to_int32_positive_(zff + sfpi::vFloat(0x3f800000));  // (bias + x * log2(a)) * N_m
 
     sfpi::vInt zii = exexp(sfpi::reinterpret<sfpi::vFloat>(z));
     sfpi::vInt zif = sfpi::exman9(sfpi::reinterpret<sfpi::vFloat>(z));
 
+    // Compute formula in Horner form
     sfpi::vFloat d1 = sfpi::vFloat(0.40196114e-7);
     sfpi::vFloat d2 = sfpi::int32_to_float(sfpi::vInt(0xf94ee7) + zif);
     sfpi::vFloat d3 = sfpi::int32_to_float(sfpi::vInt(0x560) + zif);
     d2 = d1 * d2;
     zif = _float_to_int32_positive_(d2 * d3);
 
-    // zii |= zif; // restore exponent
+    // Restore exponent
     zii = sfpi::reinterpret<sfpi::vInt>(sfpi::setexp(sfpi::reinterpret<sfpi::vFloat>(zif), 127U + zii));
 
     sfpi::vFloat y = sfpi::reinterpret<sfpi::vFloat>(zii);
@@ -86,12 +121,13 @@ sfpi_inline sfpi::vFloat _sfpu_binary_power_(sfpi::vFloat base, sfpi::vFloat pow
         sfpi::float_to_int16(pow, 0);  // int16 should be plenty, since large powers will approach 0/Inf
     sfpi::vFloat pow_rounded = sfpi::int32_to_float(pow_int, 0);
 
-    v_if(base == 0.f && pow < 0.f) {
-        y = std::numeric_limits<float>::quiet_NaN();  // negative powers of 0 are NaN, e.g. pow(0, -1.5)
+    // Division by 0 when base is 0 and pow is negative => set to NaN
+    v_if((base == 0.f || base == -0.f) && pow < 0.f) {
+        y = sfpi::vConstFloatPrgm2;  // negative powers of 0 are NaN, e.g. pow(0, -1.5)
     }
     v_endif
 
-    v_if(base < 0.0f) {  // negative base
+    v_if(base < -0.0f) {  // negative base
         // Check for integer power
         v_if(pow_rounded == pow) {
             // if pow is odd integer, set result to negative
@@ -106,14 +142,14 @@ sfpi_inline sfpi::vFloat _sfpu_binary_power_(sfpi::vFloat base, sfpi::vFloat pow
             // multiplication by NaN gives NaN.
             // Since we are going to multiply the result by `sign` to handle negative bases, we also use
             // `sign` to handle NaN results
-            y = std::numeric_limits<float>::quiet_NaN();
+            y = sfpi::vConstFloatPrgm2;  // = NaN
         }
         v_endif;
     }
     v_endif;
 
     return y;
-}  // namespace ckernel
+}
 
 template <bool APPROXIMATION_MODE, BinaryOp BINOP, int ITERATIONS = 8>
 inline void calculate_sfpu_binary(const uint dst_offset) {
@@ -138,6 +174,7 @@ inline void sfpu_binary_init() {
     if constexpr (BINOP == BinaryOp::POW) {
         sfpi::vConstFloatPrgm0 = 1.442695f;
         sfpi::vConstFloatPrgm1 = -127.0f;
+        sfpi::vConstFloatPrgm2 = std::numeric_limits<float>::quiet_NaN();
     } else {
         _sfpu_binary_init_<APPROXIMATION_MODE, BINOP>();
     }