… tests/ttnn/unit_tests/operations/eltwise/test_pow.py