### Ticket
Link to Github Issue

### Problem description

As reported by #<issue number> and #<issue number> the current implementation of the `pow()` for `(tensor, tensor)` is nto enough.

For instance, `pow(9, 2) ~= 78` instead of `81` for bfloat16 inputs.

### What's changed

This PR replaces the current implementation by a more accurate implementation. 

This implementation follows that of `power_21f` from [Simple Multiple Precision Algorithms for Exponential Functions [Tips & Tricks]](https://www.researchgate.net/publication/361844431_Simple_Multiple_Precision_Algorithms_for_Exponential_Functions_Tips_Tricks_IEEE_Signal_Processing_Magazine_394_130-137) paper from 2022 by Moroz et al. 

#### Accuracy improvement

Based on our testing, the new function is more accurate than the old one by a factor of 5. 

Notably, `new_pow(9, 2) = 80.5` which is only 1 ULP away from 81.
(Note: 1ULP can be further reduce to a perfect result by properly rounding the result to bf16 #<issue> )

To ascertain the accuracy of a large range of input, I've exhaustively tested it on every possible pair of bfloat16 inputs. 


The following heat maps show the matching accuracy results for the current `pow` function, and the proposed `pow`. 
Purple inputs = non-finite results (NaN or infinity). Darker = better. 

To mitigate the file size overhead (i.e 4 billions possible inputs would mean several gigabytes of data), I've grouped results: a single data point represents 128x128 values (i.e. values are grouped by mantissa). This is why there is a square pattern.

The value is the *maximum* error in ULP among all 128x128 results. 


<img width="400" height="357" alt="pow" src="https://github.com/user-attachments/assets/14d8284a-58e6-40b2-afc6-c448e998e04e" /> <img width="400" height="357" alt="pow21f" src="https://github.com/user-attachments/assets/2d28a168-1157-4c00-92e0-65257c3a7659" />


#### Execution time overhead

The execution time has been measured by looping 100'000 times on the LLK:
```
for (size_t i = 0; i < 100000; i++) {
    result = _sfpu_binary_power_(in0, in1);
}
```

As shown on the table, this accuracy improvement comes with a small performance penalty: it is 1% slower than the current version. 
However, the significant accuracy improvement should make this a reasonable trade-off for the time being.

For a `512x512` tensor in L1 memory:
 
| . | Execution Time [ns] | Throughput [GElem/s] | Perf vs. current |
| - | - | - | - |
| `ttnn.pow` | 12675| 20.7 | |
| this | 12800 | 20.5 | -1% |


### Future work

Due to a rounding issue, some bfloat16 results are rounded down rather than rounded to the nearest number.
For instance, `pow(9, 2) = 8.0819`, but gets truncated down to `80.5` for bfloat16 rather than to the nearest 81 (which would have been a perfect result) 

### Checklist
- [ ] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI passes
- [ ] [Blackhole Post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml) CI with demo tests passes (if applicable)
- [ ] [Model regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable)
- [ ] [Device performance regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml) CI passes (if applicable)
- [ ] (For models and ops writers) [Single-card demo tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml) CI passes (if applicable) See [recommended dev flow](https://github.com/tenstorrent/tt-metal/blob/main/models/MODEL_ADD.md#a-recommended-dev-flow-on-github-for-adding-new-models).
- [ ] [Galaxy quick](https://github.com/tenstorrent/tt-metal/actions/workflows/tg-quick-trigger.yaml) CI passes (if applicable)
- [ ] [TG demo tests, for Llama](https://github.com/tenstorrent/tt-metal/actions/workflows/tg-demo-tests.yaml) CI passes, if applicable, because of current Llama work
- [ ] (For runtime and ops writers) [T3000 unit tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-unit-tests.yaml) CI passes (if applicable, since this is run on push to main)
- [ ] (For models and ops writers) [T3000 demo tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-demo-tests.yaml) CI passes (if applicable, since this is required for release)
- [ ] (For models and ops writers) [Single-card demo tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml) CI passes (if applicable, since this is required for release)
- [ ] New/Existing tests provide coverage for changes