@kmabeeTT 2025-07-28
Hearing that there is some non-deterministic behavior coming from this op/pr caught in tt-torch and fix is being worked on?  Dropping line here so others are aware :)